refactor: move items around images to image.rs
